requirements.txt
asgiref==3.8.1
chatterbot-corpus==1.2.0
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.15.1
drf-yasg==1.21.7
environ==1.0
filelock==3.13.3
gunicorn==22.0.0
inflection==0.5.1
packaging==24.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
whitenoise==6.6.0
